Make sure there's only one .csproj per directory
Therefore split up sample folders containing multiple
projects into subfolders. This fixes the macOS build and
also some weird issues on other operating systems (such
as the debugger getting confused about which one to execute)
