+++ src/devices/Arduino/samples/ApiChecker/ArduinoCompilerSampleMethods.cs
