+++ src/devices/Arduino/samples/ApiChecker/DebugLogStream.cs
