+++ src/devices/Arduino/samples/ApiChecker/TestCases.cs
