+++ src/devices/Arduino/samples/Monitor/Arduino.Monitor.csproj
@@ -14,13 +14,12 @@
     <Compile Include="Arduino.Monitor.cs" />
   </ItemGroup>
   <ItemGroup>
-    <!--Cases used when running in VS, with the solution config either Windows-Debug or Linux-Debug (or -Release) -->
-    <ProjectReference Include="..\..\Bmxx80\Bmxx80.csproj" />
-    <ProjectReference Include="..\..\Button\Button.csproj" />
-    <ProjectReference Include="..\..\Board\Board.csproj" />
-    <ProjectReference Include="..\..\CharacterLcd\CharacterLcd.csproj" />
-    <ProjectReference Include="..\..\HardwareMonitor\HardwareMonitor.csproj" />
-    <ProjectReference Include="..\..\Mcp3xxx\Mcp3xxx.csproj" />
-    <ProjectReference Include="..\Arduino.csproj" />
+    <ProjectReference Include="..\..\..\Bmxx80\Bmxx80.csproj" />
+    <ProjectReference Include="..\..\..\Button\Button.csproj" />
+    <ProjectReference Include="..\..\..\Board\Board.csproj" />
+    <ProjectReference Include="..\..\..\CharacterLcd\CharacterLcd.csproj" />
+    <ProjectReference Include="..\..\..\HardwareMonitor\HardwareMonitor.csproj" />
+    <ProjectReference Include="..\..\..\Mcp3xxx\Mcp3xxx.csproj" />
+    <ProjectReference Include="..\..\Arduino.csproj" />
   </ItemGroup>
 </Project>
