+++ src/devices/Arduino/samples/Monitor/CharacterDisplay.cs
