+++ src/devices/Bh1745/samples/CustomConfiguration/Bh1745.CustomConfiguration.cs
