+++ src/devices/Bmxx80/samples/Bme280/Bme280.sample.csproj
@@ -1,13 +1,13 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
+<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <OutputType>Exe</OutputType>
     <TargetFramework>$(DefaultSampleTfms)</TargetFramework>
     <EnableDefaultItems>false</EnableDefaultItems>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="Program.cs" />
+    <Compile Include="Bme280.sample.cs" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="../Bh1745.csproj" />
+    <ProjectReference Include="../../Bmxx80.csproj" />
   </ItemGroup>
 </Project>
