+++ src/devices/Bmxx80/samples/Bme680/Bme680.sample.csproj
@@ -5,10 +5,9 @@
     <EnableDefaultItems>false</EnableDefaultItems>
   </PropertyGroup>
   <ItemGroup>
-    <ProjectReference Include="../Seesaw.csproj" />
-    <Compile Include="Seesaw.Sample.Capabilities.cs" />
+    <Compile Include="Bme680.sample.cs" />
   </ItemGroup>
   <ItemGroup>
-    <None Include="README.md" />
+    <ProjectReference Include="../../Bmxx80.csproj" />
   </ItemGroup>
 </Project>
