--- src/devices/Bmxx80/samples/Bmp280.sample.csproj
@@ -1,17 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>$(DefaultSampleTfms)</TargetFramework>
-    <EnableDefaultItems>false</EnableDefaultItems>
-  </PropertyGroup>
-  <ItemGroup>
-    <None Remove="rpi-bmp280_i2c.fzz" />
-    <None Remove="rpi-bmp280_i2c.png" />
-  </ItemGroup>
-  <ItemGroup>
-    <Compile Include="Bmp280.sample.cs" />
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\Bmxx80.csproj" />
-  </ItemGroup>
-</Project>
