+++ src/devices/Bmxx80/samples/Bmp280/Bmp280.sample.csproj
@@ -5,9 +5,9 @@
     <EnableDefaultItems>false</EnableDefaultItems>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="Program.CustomConfiguration.cs" />
+    <Compile Include="Bmp280.sample.cs" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="../Bh1745.csproj" />
+    <ProjectReference Include="../../Bmxx80.csproj" />
   </ItemGroup>
 </Project>
