+++ src/devices/Seesaw/samples/BlinkingLights/Seesaw.Sample.BlinkingLights.csproj
@@ -1,0 +1,13 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>$(DefaultSampleTfms)</TargetFramework>
+    <EnableDefaultItems>false</EnableDefaultItems>
+  </PropertyGroup>
+  <ItemGroup>
+    <ProjectReference Include="../../Seesaw.csproj" />
+    <Compile Include="Seesaw.Sample.BlinkingLights.cs" />
+    <Compile Include="../../../../../samples/led-more-blinking-lights/TimeEnvelope.cs" />
+    <Compile Include="Volume.cs" />
+  </ItemGroup>
+</Project>
