+++ src/devices/Seesaw/samples/Capabilities/Seesaw.Sample.Capabilities.cs
