+++ src/devices/Seesaw/samples/Encoder/Seesaw.Sample.Encoder.csproj
@@ -5,13 +5,9 @@
     <EnableDefaultItems>false</EnableDefaultItems>
   </PropertyGroup>
   <ItemGroup>
-    <None Remove="rpi-bmp280_i2c.fzz" />
-    <None Remove="rpi-bmp280_i2c.png" />
+    <Compile Include="Seesaw.Sample.Encoder.cs" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Bme680.sample.cs" />
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\Bmxx80.csproj" />
+    <ProjectReference Include="../../Seesaw.csproj" />
   </ItemGroup>
 </Project>
