+++ src/devices/Seesaw/samples/SoilSensor/Seesaw.Sample.SoilSensor.csproj
